clk: canaan: Add clock driver for Canaan K230
This patch provides basic support for the K230 clock, does not cover all
clocks.

The clock tree of the K230 SoC consists of PLLs and sysclk.
Since the PLLs are not directly used by clock consumers, but are instead
used by sysclk, support can be implemented with just a single
compatible.

Signed-off-by: Xukai Wang <kingxukai@zohomail.com>
Signed-off-by: Troy Mitchell <TroyMitchell988@gmail.com>